backed out previous change which checked for RRULE's - it didn't make sense as it was checking for RRULE on a calendar as opposed to the events inside.
Also, added behavior in InstanceList which converts floating DateTimes to UTC for comparisons. 
Without this, cosmo is rather unpredictable as it just converts to whatever the cosmo servers local time is.

Of course, the 'real' way to deal with this is by making the calendar queries take the calendar collections and the queries
time-zone into account.

git-svn-id: http://svn.osafoundation.org/server/ical4j/trunk@1285 5fd0b11a-f2f7-0310-948e-c8f80ebed640